Trait oni::Socket

pub trait Socket: Sized {
    fn bind(addr: SocketAddr) -> Result<Self>;

    fn local_addr(&self) -> Result<SocketAddr>;

    fn recv_from(&self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>;

    fn send_to(&self, buf: &[u8], addr: SocketAddr) -> Result<usize>;

    fn connect(&self, addr: SocketAddr) -> Result<()>;

    fn send(&self, buf: &[u8]) -> Result<usize>;

    fn recv(&self, buf: &mut [u8]) -> Result<usize>;

    fn set_nonblocking(&self, nonblocking: bool) -> Result<()>;
}

Required Methods

fn bind(addr: SocketAddr) -> Result<Self>

Creates a socket from the given address.

fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address that this socket was created from.

fn recv_from(&self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>

Receives a single datagram message on the socket. On success, returns the number of bytes read and the origin.

Simulated socket

The function must be called with valid byte array buf of sufficient size to hold the message bytes. If a message is too long to fit in the supplied buffer, excess bytes may be discarded.

fn send_to(&self, buf: &[u8], addr: SocketAddr) -> Result<usize>

Sends data on the socket to the given address. On success, returns the number of bytes written.

Simulated socket

This will return an error when the length of buf is greater than MTU.

fn connect(&self, addr: SocketAddr) -> Result<()>

fn send(&self, buf: &[u8]) -> Result<usize>

fn recv(&self, buf: &mut [u8]) -> Result<usize>

fn set_nonblocking(&self, nonblocking: bool) -> Result<()>

Simulated socket

Does nothing.